disable bootdelay to not interrupt the recovery image
By default u-boot will wait for a keystroke before starting to run
the boot command. If a device is connected to the serial console and
transmitting data the recovery process is also interrupted, which is
not the intention. This behavior is changed to always ignore incoming
characters and drop to the u-boot shell if the left function key is
pressed.

NOTE:
- This does need an u-boot update since the preload variable is only
evaluated when the CONFIG_PRELOAD is defined (which was not the case).
- This does need an env change as well, since the config only sets the
default env value.

Signed-off-by: Jeroen Hofstee <jhofstee@victronenergy.com>